Issue #202: Reduce the file length of InputLocalVariableCheckSwitchExpressions
# Conflicts:
#	config/checkstyle-resources-suppressions.xml
