+++ config/checkstyle-resources-suppressions.xml
@@ -7,8 +7,6 @@
 <suppressions>
     <!-- Until https://github.com/checkstyle/checkstyle-openrewrite-recipes/issues/202 -->
     <suppress checks="FileLength"
-              files="[\\/]src[\\/]test[\\/]resources[\\/]org[\\/]checkstyle[\\/]autofix[\\/]recipe[\\/]finallocalvariable[\\/]localvariablecheckswitchexpressions[\\/]InputLocalVariableCheckSwitchExpressions\.java"/>
-    <suppress checks="FileLength"
               files="[\\/]src[\\/]test[\\/]resources[\\/]org[\\/]checkstyle[\\/]autofix[\\/]recipe[\\/]useenhancedswitch[\\/]mutationtests14[\\/]InputMutationTests14\.java"/>
     <suppress checks="FileLength"
               files="[\\/]src[\\/]test[\\/]resources[\\/]org[\\/]checkstyle[\\/]autofix[\\/]recipe[\\/]useenhancedswitch[\\/]mutationtests15[\\/]InputMutationTests15\.java"/>
