+++ src/test/resources/org/checkstyle/autofix/recipe/finallocalvariable/localvariablecheckswitchexpressions/DiffLocalVariableCheckSwitchExpressions.diff
@@ -22,21 +22,3 @@
  
          final int a = (int) Math.random();
          final int b = (int) Math.random();
-@@ -88,7 +88,7 @@
- 
-     void foo3() throws Exception {
- 
--        Exception e; // violation, "Variable 'e' should be declared final"
-+        final Exception e;
- 
-         final int a = (int) Math.random();
-         final int b = (int) Math.random();
-@@ -122,7 +122,7 @@
- 
-     void foo4() throws Exception {
- 
--        Exception e; // violation, "Variable 'e' should be declared final"
-+        final Exception e;
- 
-         final int a = (int) Math.random();
-         final int b = (int) Math.random();
